requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
markdown==3.5.1
gunicorn==21.2.0